% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mls_data.R
\name{espn_mls_game_all}
\alias{espn_mls_game_all}
\title{Get ESPN's MLS game data (play-by-play, team and player box)}
\usage{
espn_mls_game_all(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
A named list of dataframes: Plays, Team, Player
}
\description{
Get ESPN's MLS game data (play-by-play, team and player box)
}
\examples{
\donttest{
  try(espn_mls_game_all(game_id = 401244185))
}
}
\author{
Saiem Gilani
}
\keyword{Game}
\keyword{MLS}
