% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mls_data.R
\name{espn_mls_pbp}
\alias{espn_mls_pbp}
\title{Get ESPN's MLS play by play data}
\usage{
espn_mls_pbp(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
Returns a play-by-play data frame
}
\description{
Get ESPN's MLS play by play data
}
\examples{

\donttest{
  try(espn_mls_pbp(game_id = 401244185))
}
}
\author{
Saiem Gilani
}
\keyword{MLS}
\keyword{PBP}
