% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mls_data.R
\name{espn_mls_scoreboard}
\alias{espn_mls_scoreboard}
\title{Get MLS schedule for a specific year/date from ESPN's API}
\usage{
espn_mls_scoreboard(season)
}
\arguments{
\item{season}{Either numeric or character}
}
\value{
Returns a tibble
}
\description{
Get MLS schedule for a specific year/date from ESPN's API
}
\examples{
# Get schedule from date 2020-08-29
\donttest{
  try(espn_mls_scoreboard (season = "20200829"))
}
}
\author{
Saiem Gilani.
}
