% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mls_data.R
\name{espn_mls_standings}
\alias{espn_mls_standings}
\title{Get ESPN MLS Standings}
\usage{
espn_mls_standings(year)
}
\arguments{
\item{year}{Either numeric or character (YYYY)}
}
\value{
Returns a tibble
}
\description{
Get ESPN MLS Standings
}
\examples{
\donttest{
  try(espn_mls_standings(year = 2021))
}
}
\keyword{MLS}
\keyword{Standings}
