% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_mls_data.R
\name{espn_mls_teams}
\alias{espn_mls_teams}
\title{Get ESPN's MLS team names and ids}
\usage{
espn_mls_teams()
}
\value{
Returns a tibble
}
\description{
Get ESPN's MLS team names and ids
}
\examples{
\donttest{
  try(espn_mls_teams())
}
}
\author{
Saiem Gilani
}
\keyword{MLS}
\keyword{Teams}
