% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_nwsl_data.R
\name{espn_nwsl_game_all}
\alias{espn_nwsl_game_all}
\title{Get ESPN's NWSL game data (play-by-play, team)}
\usage{
espn_nwsl_game_all(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
A named list of dataframes: Plays, Team
}
\description{
Get ESPN's NWSL game data (play-by-play, team)
}
\examples{
\donttest{
  try(espn_nwsl_game_all(game_id = 601833))
}
}
\author{
Saiem Gilani
}
\keyword{Game}
\keyword{NWSL}
