% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_nwsl_data.R
\name{espn_nwsl_pbp}
\alias{espn_nwsl_pbp}
\title{Get ESPN's NWSL play by play data}
\usage{
espn_nwsl_pbp(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
Returns a play-by-play data frame
}
\description{
Get ESPN's NWSL play by play data
}
\examples{

\donttest{
  try(espn_nwsl_pbp(game_id = 401244185))
}
}
\author{
Saiem Gilani
}
\keyword{NWSL}
\keyword{PBP}
