% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_nwsl_data.R
\name{espn_nwsl_player_box}
\alias{espn_nwsl_player_box}
\title{Get ESPN's NWSL player box data}
\usage{
espn_nwsl_player_box(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
Returns a player boxscore data frame
}
\description{
Get ESPN's NWSL player box data
}
\examples{
\donttest{
  try(espn_nwsl_player_box(game_id = 401244185))
}

}
\author{
Saiem Gilani
}
\keyword{Box}
\keyword{NWSL}
\keyword{Player}
