% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_nwsl_data.R
\name{espn_nwsl_scoreboard}
\alias{espn_nwsl_scoreboard}
\title{Get NWSL schedule for a specific year/date from ESPN's API}
\usage{
espn_nwsl_scoreboard(season)
}
\arguments{
\item{season}{Either numeric or character}
}
\value{
Returns a tibble
}
\description{
Get NWSL schedule for a specific year/date from ESPN's API
}
\examples{
# Get schedule from date 2020-08-29
\donttest{
  try(espn_nwsl_scoreboard (season = "20200829"))
}
}
\author{
Saiem Gilani.
}
