% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_nwsl_data.R
\name{espn_nwsl_standings}
\alias{espn_nwsl_standings}
\title{Get ESPN NWSL Standings}
\usage{
espn_nwsl_standings(year)
}
\arguments{
\item{year}{Either numeric or character (YYYY)}
}
\value{
Returns a tibble
}
\description{
Get ESPN NWSL Standings
}
\examples{
\donttest{
  try(espn_nwsl_standings(year = 2021))
}
}
\author{
Geoff Hutchinson
}
\keyword{NWSL}
\keyword{Standings}
