% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_nwsl_data.R
\name{espn_nwsl_team_box}
\alias{espn_nwsl_team_box}
\title{Get ESPN's NWSL team box data}
\usage{
espn_nwsl_team_box(game_id)
}
\arguments{
\item{game_id}{Game ID}
}
\value{
Returns a team boxscore data frame
}
\description{
Get ESPN's NWSL team box data
}
\examples{

\donttest{
  try(espn_nwsl_team_box(game_id = 401244185))
}
}
\author{
Saiem Gilani
}
\keyword{Box}
\keyword{NWSL}
\keyword{Team}
