% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_nwsl_data.R
\name{espn_nwsl_teams}
\alias{espn_nwsl_teams}
\title{Get ESPN's NWSL team names and ids}
\usage{
espn_nwsl_teams()
}
\value{
Returns a tibble
}
\description{
Get ESPN's NWSL team names and ids
}
\examples{
\donttest{
  try(espn_nwsl_teams())
}
}
\author{
Saiem Gilani
}
\keyword{NWSL}
\keyword{Teams}
