% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwsl_pbp.R
\name{load_nwsl_schedule}
\alias{load_nwsl_schedule}
\title{\strong{Load usfootballR NWSL schedules}}
\usage{
load_nwsl_schedule(
  seasons = most_recent_nwsl_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)
}
\arguments{
\item{seasons}{A vector of 4-digit years associated with given NWSL seasons. (Min: 2002)}

\item{...}{Additional arguments passed to an underlying function that writes
the season data into a database (used by \code{update_nwsl_db()}).}

\item{dbConnection}{A \code{DBIConnection} object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{tablename}{The name of the schedule data table within the database}
}
\value{
Returns a tibble
}
\description{
helper that loads multiple seasons from the data repo either into memory
or writes it into a db using some forwarded arguments in the dots
}
\examples{
\donttest{
load_nwsl_schedule(2021)
}
}
